fix links to images;
