+++ cs/TagsCloudVisualization/readme.md
@@ -1,2 +1,2 @@
-﻿![Example of not so cloud](.\image.png)
-![Example of not so cloud](.\circular.png)
+﻿![Example of not so cloud](image.png)
+![Example of not so cloud](circular.png)
